Add files via upload
This is one of my First logo design for a cofee shop, when I was working as a freelancer. The photo was taken from screen of  my laptop, and then edited, please forge it't apearence. I have done research about it and made the best logo according to their desire.